moved Door43DataFetcher.js and ViewEventListeners.js to new js folder and added require for usfm-parser npm module
